server/requirements.txt
pipecat-ai[google,webrtc,daily]
pipecat-ai-small-webrtc-prebuilt
python-dotenv
fastapi
uvicorn[standard]
aiohttp
requests
python-multipart
Pillow
numpy
scipy
